Rename folder with proper capitalization
